Moved SideWindow-related code from Eagle to SideWindow.ts
